starts fixing shit
