+++ src/Components/Buttons/BackButton.tsx
@@ -2,7 +2,6 @@ import { Button } from 'antd';
 import { useNavigate } from 'react-router-dom';
 import { useContext } from 'react';
 import { TranslationContext } from '../../Contexts/TranslationContext';
-import NavigationButtons from './NavigationButtons';
 
 function BackButton({ scrolling = false }: any) {
   const { t } = useContext(TranslationContext);
